+++ lib/widgets/keyboard_listener.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 enum InputShortcut { CUT, COPY, PASTE, SELECT_ALL }
@@ -61,6 +62,11 @@ class KeyboardListener {
   KeyboardListener(this.onCursorMove, this.onShortcut, this.onDelete);
 
   bool handleRawKeyEvent(RawKeyEvent event) {
+    if (kIsWeb) {
+      // On web platform, we should ignore the key because it's processed already.
+      return false;
+    }
+
     if (event is! RawKeyDownEvent) {
       return false;
     }